Move picard/util/remotecommands.py to picard/remotecommands/__init__.py
